fix(determinism): address the determinism in the gnark circuit
